Vain Riser Project Support (#109)
* Async Future Bug Fixed

* Bug Fixed

* Bug Fixed

* Font Update

* ScriptPostProcessor

* Bug fixed

* Auto Wrap Bitmap To 4

* Group::FindFirst

* Bug Fixed

* Update

* Touch Support (#113)

* Touch Support

* Bug fixed.

* Touch

* BUG  FIXED.

* bug fixed.

* bug fixed

* Bug Fixed

* Graphics Update

* Bug Fixed

* Fuck you junk aji microsoft,your shit api make my code shabi, FUCK U!

* Bug Fixed

* Bug Fixed

* Bug fixed

* Bug fixed

* Batch

* Bug fixed.

* Lip Sync Bug Fixed.

* Bug fixed

* Menu Bug Fixed

* Bug fixed

* Snowing::CallAssetBuilder

* bug fixed

* bug fixed

* bug fixed

* bug fixed

* bug fixed:

* bug fixed

* bug fixed.

* Revert "bug fixed."

This reverts commit 2aa1dc0bcf80c91c366ec7afe9c78727a73b99bd.

* Revert "bug fixed"

This reverts commit 63935704bf572943170fdaff542fa00c1203ff81.

* Revert "bug fixed:"

This reverts commit 3e50a745008edf4bf9bde03ff4d2145e3b2e8b2c.

* Revert "bug fixed"

This reverts commit 58253266ee48905337bf2bdc1130dfce4286496a.

* Revert "bug fixed"

This reverts commit 2d02120bae5536d80e42e463000c95761dbb1a2f.

* Revert "bug fixed"

This reverts commit dfa4fbcd253d9ba58cc4fdfff3484d5b8b0e209a.

* Fyee Project (#121)

* Fyee Init

* Demo

* SuperLooper is finished

* BreakOnNextLoop

* Schedule Functions

* Finished

* bug fixed

* eaiser to use.

* remove bgm player

* bug fixed

* bug fixed

* bug fixed

* bug fixed

* Optimized

* Snowing::Graphics::PixelFormat::R8

* Default to DirectX 11

* bug fixed

* Optimized

* bug fixed

* Debug mode bug fixed

* may bug fixed

* bug fixed

* bug fixed

* Update To MSVC V142 Win10 1903 SDK

* 对win7兼容性提升

* 兼容性问题

* BUG FIXED.

* Down to 8.1 SDK

* bug fixed

* XAudio2.7

* remove some file

* Bug fixed

* windows 7 support

* AssetBuilder Job GenerateVersionText

* CV API Update

* Engine Update

* Live2D LipSync bug fixed.

* Bug Fixed

* 鼠标滚轮输出精确数值，支持横向滚动滚轮

* AVG Input不再将鼠标滚轮模拟为按键设备

* bug fixed

* Disable IME

* Menu bug fixed

* Bug fixed

* Fyee bug fixed.

* DebugMenuClickable

* Update Abort Method

* Bug fixed

* Fyee Update

* Fyee Bug fixed

* bug fixed

* Sign Tool Used

* bug fixed

* AssetBuilder update

* Bug fixed

* AssetsBuilder supports AES Encryptor

* Bug fixed

* Remove AES

* Bug fixed

* Bug fixed

* Bug fixed

* Bug fixed

* Win7 Support

* MP Compile

* Bug fixed

* CVPlayer::IsPlaying()

* TouchScreen Touch Point Position Bug fixed

* Bug fixed

* Update font

* Bug fixed

* bug fixed

* Live2D::Expression::SetWeight

* Expression Bug fixed

* Live2D Wind Bug Fixed

* Bug fixed

* Asset Builder Bug fixed

* Rebuild when script file is newer

* Bug fixed

* Clearly Debug Display

* Bug fixed

* Fyee BGM

* Reset Fyee with fadeout time

* ClearCoda

* Bug fixed

* Bug fixed

* Bug fixed

* Bug fixed

* Bug fixed

* Bug fixed

* Live2D Updated

* Live 2D Updated

* Bug fixed

* Bug fixed

* bug fixed

* Strong Typed Shader Reflection Object

* Strong Typed Reflection

* Bug fixed

* Bug fixed

* Tween::Add

* Optimized

* Bake (#123)

* 废弃AssetsBuilder，使用Bake构建系统取代

* Bug fixed

* Bug fixed

* Bug fixed

* Bug fixed

* Debugging Steam Proton

* Debugging Proton

* Debugging Proton

* Debug for Proton

* Debugging Proton

* Debugging Proton

* Debugging Proton

* Debugging Proton

* Debugging Proton

* Bug fixed

* bug fixed

* Update XAudio to 2.9 (#125)

* Update to XAudio 2.9

* Bug fixed

* Tested

* bug fixed

* Bug fixed

* Pack to a class

* 全屏模式下鼠标Y轴偏移Bug修正

* Bug fixed

* Bug fixed

* Bug fixed

* Bug fixed

* Vec::Cast

* Bug fixed

* Bug fixed

* Sizable window

* Bug fixed

* Optimized

* bug fixed

* Add ClearState()

* Bug fixed

* Bug fixed

* Bug fixed

* Bug fixed

* SetDebuggingScreen

* Draw Lines

* SetIcon

* CSVParser(std::basic_string_view<CharType>)

* SetTexture

* Page Down / Page Up key

* Bug fixed

* Bug fixed

* Alt+Tab Bug fixed

* Bug fixed

* bug fixed

* Bug fixed

* bug fixed

* Update TexConv

* BinPack

* Bug fixed

* Bug fixed

* bug fixed

* More Error Information

* bug fixed

* Bug fixed

* Bug fixed

* Windows 7 support

* Disable Alt+Enter

+++ Core/Window.h
@@ -21,12 +21,12 @@ namespace Snowing::Graphics
 			TImpl::Get().SetWindowed(b);
 		}
 
-		static Math::Vec2<int> GetSize()
+		static Math::Vec2<size_t> GetSize()
 		{
 			return TImpl::Get().GetSize();
 		}
 
-		static void Resize(Math::Vec2<int> size)
+		static void Resize(Math::Vec2<size_t> size)
 		{
 			TImpl::Get().Resize(size);
 		}
